feat: added svelte hash router and updated all app dependencies including svelte4
