+++ package.json
@@ -1,60 +1,61 @@
 {
-  "name": "office-addin-taskpane-svelte",
-  "private": true,
-  "version": "1.0.0",
-  "type": "module",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/krmanik/Office-Addin-TaskPane-Svelte.git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "lint": "office-addin-lint check",
-    "lint:fix": "office-addin-lint fix",
-    "prettier": "office-addin-lint prettier",
-    "start": "office-addin-debugging start ./manifests/manifest.xml",
-    "start:desktop": "office-addin-debugging start ./manifests/manifest.xml desktop",
-    "start:web": "office-addin-debugging start ./manifests/manifest.xml web",
-    "start:excel": "office-addin-debugging start ./manifests/manifest.excel.xml",
-    "start:onenote": "office-addin-debugging start ./manifests/manifest.onenote.xml",
-    "start:powerpoint": "office-addin-debugging start ./manifests/manifest.powerpoint.xml",
-    "start:project": "office-addin-debugging start ./manifests/manifest.project.xml",
-    "start:word": "office-addin-debugging start ./manifests/manifest.word.xml",
-    "stop": "office-addin-debugging stop ./manifests/manifest.xml",
-    "validate": "office-addin-manifest validate ./manifests/manifest.xml"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.21.0",
-    "@babel/eslint-parser": "^7.19.1",
-    "@babel/preset-typescript": "^7.21.0",
-    "@microsoft/office-js": "^1.1.84",
-    "@sveltejs/vite-plugin-svelte": "^2.0.2",
-    "@tsconfig/svelte": "^3.0.0",
-    "@types/office-js": "^1.0.318",
-    "autoprefixer": "^10.4.19",
-    "eslint-plugin-office-addins": "^2.1.5",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "office-addin-debugging": "^5.0.4",
-    "office-addin-lint": "^2.2.5",
-    "office-addin-prettier-config": "^1.2.0",
-    "postcss": "^8.4.38",
-    "svelte": "^3.55.1",
-    "svelte-check": "^2.10.3",
-    "tailwindcss": "^3.4.3",
-    "ts-node": "^10.9.1",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.3",
-    "vite": "^4.1.0",
-    "vite-plugin-html": "^3.2.0"
-  },
-  "dependencies": {
-    "@fluentui/web-components": "^2.5.12",
-    "@microsoft/fast-element": "^1.11.0",
-    "lodash-es": "^4.17.21"
-  },
-  "office-addin-prettier-config": "^1.2.0"
+	"name": "office-addin-taskpane-svelte",
+	"private": true,
+	"version": "1.0.0",
+	"type": "module",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/krmanik/Office-Addin-TaskPane-Svelte.git"
+	},
+	"license": "MIT",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build",
+		"preview": "vite preview",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"lint": "office-addin-lint check",
+		"lint:fix": "office-addin-lint fix",
+		"prettier": "office-addin-lint prettier",
+		"start": "office-addin-debugging start ./manifests/manifest.xml",
+		"start:desktop": "office-addin-debugging start ./manifests/manifest.xml desktop",
+		"start:web": "office-addin-debugging start ./manifests/manifest.xml web",
+		"start:excel": "office-addin-debugging start ./manifests/manifest.excel.xml",
+		"start:onenote": "office-addin-debugging start ./manifests/manifest.onenote.xml",
+		"start:powerpoint": "office-addin-debugging start ./manifests/manifest.powerpoint.xml",
+		"start:project": "office-addin-debugging start ./manifests/manifest.project.xml",
+		"start:word": "office-addin-debugging start ./manifests/manifest.word.xml",
+		"stop": "office-addin-debugging stop ./manifests/manifest.xml",
+		"validate": "office-addin-manifest validate ./manifests/manifest.xml"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.24.5",
+		"@babel/eslint-parser": "^7.24.5",
+		"@babel/preset-typescript": "^7.24.1",
+		"@microsoft/office-js": "^1.1.93",
+		"@sveltejs/vite-plugin-svelte": "^3.1.0",
+		"@tsconfig/svelte": "^5.0.4",
+		"@types/office-js": "^1.0.390",
+		"autoprefixer": "^10.4.19",
+		"eslint-plugin-office-addins": "^3.0.1",
+		"eslint-plugin-svelte": "^2.39.0",
+		"office-addin-debugging": "^5.1.2",
+		"office-addin-lint": "^2.3.2",
+		"office-addin-prettier-config": "^1.2.0",
+		"postcss": "^8.4.38",
+		"svelte": "^4.2.17",
+		"svelte-check": "^3.7.1",
+		"tailwindcss": "^3.4.3",
+		"ts-node": "^10.9.2",
+		"tslib": "^2.6.2",
+		"typescript": "^5.4.5",
+		"vite": "^5.2.11",
+		"vite-plugin-html": "^3.2.2"
+	},
+	"dependencies": {
+		"@fluentui/web-components": "^2.6.1",
+		"@microsoft/fast-element": "^1.13.0",
+		"lodash-es": "^4.17.21",
+		"svelte-spa-router": "^4.0.1"
+	},
+	"office-addin-prettier-config": "^1.2.0"
 }
