+++ src/layouts/AppLayout.svelte
@@ -1,0 +1,10 @@
+<script>
+import Navbar from "../components/Navbar.svelte";
+</script>
+
+<Navbar />
+
+<main>
+    <slot />
+</main>
+
