+++ src/main.ts
@@ -1,8 +1,8 @@
-import "./app.css";
 import App from "./App.svelte";
+import "./app.css";
 
 const app = new App({
-  target: document.getElementById("app"),
+	target: document.getElementById("app"),
 });
 
 export default app;
