+++ src/routes/Login.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+</script>
+
+<div class="flex flex-col items-center justify-center h-screen">
+  <h1 class="text-4xl font-bold">Login</h1>
+</div>
+
